Fix compile errors when building under mingw64

+++ zbar/decoder.c
@@ -25,6 +25,7 @@
 #include <stdlib.h>     /* malloc, calloc, free */
 #include <stdio.h>      /* snprintf */
 #include <string.h>     /* memset, strlen */
+#include <stdint.h>
 
 #include <zbar.h>
 
@@ -109,7 +110,7 @@ void zbar_decoder_destroy (zbar_decoder_t *dcode)
 
 void zbar_decoder_reset (zbar_decoder_t *dcode)
 {
-    memset(dcode, 0, (long)&dcode->buf_alloc - (long)dcode);
+    memset(dcode, 0, (uintptr_t)&dcode->buf_alloc - (uintptr_t)dcode);
 #ifdef ENABLE_EAN
     ean_reset(&dcode->ean);
 #endif
